Remove chat_template file to avoid circular import
